requirements.txt
#
fastapi==0.54.1
snowflake-sqlalchemy==1.2.3

##### Core scientific packages
jupyter==1.0.0
matplotlib==3.1.2
numpy==1.17.3
pandas==0.25.3
scipy==1.3.1

# Efficient jobs (caching, parallelism, persistence)
joblib==0.14.0

# Easy http requests
requests==2.22.0

##### Machine Learning packages
scikit-learn==0.22

# Optional: the XGBoost library is only used in the ensemble learning chapter.
xgboost==0.90

# Optional: tqdm displays nice progress bars, ipywidgets for tqdm's notebook support
tqdm==4.40.0
ipywidgets==7.5.1

# pandas support for snowflake connector
snowflake-connector-python[pandas]==2.2.5

streamlit==0.59.0